+++ Text-ConsoleRPG/Dungeon/Altarevent.h
@@ -3,8 +3,8 @@
 class Player;
 
 struct AltarResult {
-    bool blessed = false;      // true = 축복(회복), false = 저주(피해)
-    unsigned short amount = 0;
+    bool blessed = false;   // true = 축복(회복), false = 저주(피해)
+    int amount = 0;
 };
 
 class AltarEvent {
